+++ CHANGELOG.md
@@ -20,3 +20,4 @@
 - [Feature: Add router actuate endpoint](https://github.com/Tencent/spring-cloud-tencent/pull/464)
 - [feat:rename the configuration of reporter.](https://github.com/Tencent/spring-cloud-tencent/pull/465)
 - [docs:update mvnw.](https://github.com/Tencent/spring-cloud-tencent/pull/466)
+- [Feature: add config refresh type switch & catch exception for traffic staining.](https://github.com/Tencent/spring-cloud-tencent/pull/469)